+++ ci/checks/enforce-mocking-only-whitelisted-methods.sh
@@ -13,7 +13,7 @@ git_machete.client.MacheteClient.is_stdout_a_tty
 git_machete.git_operations.GitContext.fetch_remote
 git_machete.github.GitHubClient.MAX_PULLS_PER_PAGE_COUNT
 git_machete.github.GitHubToken.for_domain
-git_machete.github.RemoteAndOrganizationAndRepository.from_url
+git_machete.github.OrganizationAndRepository.from_url
 git_machete.utils._popen_cmd
 git_machete.utils._run_cmd
 git_machete.utils.find_executable