+++ RELEASE_NOTES.md
@@ -2,6 +2,8 @@
 
 ## New in git-machete 3.18.1
 
+- fixed: `machete.github.remote` git config key can be specified independently from `machete.github.organization` and `machete.github.repository`
+
 ## New in git-machete 3.18.0
 
 - added: `git machete completion bash|fish|zsh` command